+++ src/locales/es/translation.json
@@ -26,7 +26,7 @@
   "pages": {
     "home": {
       "name": "GERARD HEREDIA",
-      "role": "Desarrolador Full Stack"
+      "role": "Desarrollador Full Stack"
     }
   }
 }
