+++ src/pages/Home/Home.tsx
@@ -4,7 +4,7 @@ export default function Home() {
   const { t } = useTranslation();
 
   return (
-    <div className="flex-grow mx-auto max-w-screen-xl bg-tertiary">
+    <div className="flex-grow mx-auto max-w-screen-xl">
       <div>
         <h2 className="font-medium text-6xl">{t("pages.home.name")}</h2>
         <p className="text-4xl">{t("pages.home.role")}</p>
